+++ frontend/src/pages/notebookController/screens/server/SizeSelectField.tsx
@@ -1,6 +1,7 @@
 import * as React from 'react';
 import { FormGroup, Select, SelectOption } from '@patternfly/react-core';
 import AppContext from '../../../../app/AppContext';
+import { NotebookControllerContext } from '../../NotebookControllerContext';
 
 type SizeSelectFieldProps = {
   value: string;
@@ -10,7 +11,24 @@ type SizeSelectFieldProps = {
 const SizeSelectField: React.FC<SizeSelectFieldProps> = ({ value, setValue }) => {
   const [sizeDropdownOpen, setSizeDropdownOpen] = React.useState<boolean>(false);
   const { dashboardConfig } = React.useContext(AppContext);
-  console.log(value);
+  const { currentUserState } = React.useContext(NotebookControllerContext);
+
+  React.useEffect(() => {
+    if (dashboardConfig?.spec.notebookSizes) {
+      if (currentUserState?.lastSelectedSize) {
+        const size = dashboardConfig.spec.notebookSizes.find(
+          (notebookSize) => notebookSize.name === currentUserState.lastSelectedSize,
+        );
+        if (size) {
+          setValue(size.name);
+        } else {
+          setValue(dashboardConfig.spec.notebookSizes[0].name);
+        }
+      } else {
+        setValue(dashboardConfig.spec.notebookSizes[0].name);
+      }
+    }
+  }, [dashboardConfig, currentUserState, setValue]);
 
   const sizeOptions = React.useMemo(() => {
     const sizes = dashboardConfig?.spec?.notebookSizes;
@@ -21,12 +39,10 @@ const SizeSelectField: React.FC<SizeSelectFieldProps> = ({ value, setValue }) =>
     return sizes.map((size) => {
       const name = size.name;
       const desc =
-        name === 'Default'
-          ? 'Resources set based on administrator configurations'
-          : `Limits: ${size?.resources?.limits?.cpu || '??'} CPU, ` +
-            `${size?.resources?.limits?.memory || '??'} Memory ` +
-            `Requests: ${size?.resources?.requests?.cpu || '??'} CPU, ` +
-            `${size?.resources?.requests?.memory || '??'} Memory`;
+        `Limits: ${size?.resources?.limits?.cpu || '??'} CPU, ` +
+        `${size?.resources?.limits?.memory || '??'} Memory ` +
+        `Requests: ${size?.resources?.requests?.cpu || '??'} CPU, ` +
+        `${size?.resources?.requests?.memory || '??'} Memory`;
       return <SelectOption key={name} value={name} description={desc} />;
     });
   }, [dashboardConfig]);
